+++ packages/varlet-ui/src/button/example/index.vue
@@ -58,6 +58,7 @@
   <app-type>{{ pack.event }}</app-type>
   <var-button type="success" @click="handleClick">{{ pack.click }}</var-button>
   <var-button type="success" @touchstart="handleTouchstart">{{ pack.touchstart }}</var-button>
+  <var-button type="success" @click="handleAutoLoadingClick" auto-loading>{{ pack.autoLoading }}</var-button>
 </template>
 
 <script>
@@ -86,6 +87,12 @@ export default {
       Snackbar.success(pack.value.touchstartSuccess)
     }
 
+    const handleAutoLoadingClick = () => {
+      return new Promise((resolve) => {
+        setTimeout(resolve, 2000)
+      })
+    }
+
     watchLang(use)
     watchDarkMode(dark)
 
@@ -103,6 +110,7 @@ export default {
       pack,
       handleClick,
       handleTouchstart,
+      handleAutoLoadingClick,
     }
   },
 }